Получение Userа и запоминание в userStore
