--- src/Entity/Profile.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-class Profile {
-    middle_name: string = "";
-}
-
-export default Profile;
