--- src/ViewModel/Interface/IAuthUser.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-
-interface IAuthUser {
-    username: string;
-    password: string;
-}
-
-export default IAuthUser;
